Quite a large problem was found as code I have written before from whiteboard onto VS Code which logically works but was not syntactically 100% correct;
I had to fix broken classes; Implementing in a proper way using Interfaces and Classes.

Question 3; that was a partial cause of the issue has been fixed, after using an Vector that I programmed instead of the ArrayList in System.Collections.

Q3 Now returns the right List of LinkedLists.

Lots of nullexceptionpointers, maybe for the future use a terminiating object or some sorts.